+++ src/tools/google/index.ts
@@ -1,3 +1,3 @@
 export * as ServiceAccount from './jwt-auth.js';
 export * as UniversalAnalytics from './universal-analytics.js';
-export * as PageSpeed from './pagespeed.js';
+export * from './pagespeed.js';